refactoring
extracted ACITWeapon
