+++ RaCTrainer/racman.csproj
@@ -130,6 +130,7 @@
     </Compile>
     <Compile Include="Memory\RPCS3.cs" />
     <Compile Include="offsets\ACIT\ACITAddresses.cs" />
+    <Compile Include="offsets\ACIT\ACITWeapon.cs" />
     <Compile Include="offsets\ACIT\ACITWeaponFactory.cs" />
     <Compile Include="RAC2\RC2Unlocks.cs">
       <SubType>Form</SubType>
